+++ src/components/admin/deleteBtn.js
@@ -1,17 +1,20 @@
+/* eslint-disable no-restricted-globals */
 import React from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { deleteReq } from '../../shared/Fetch';
 
-const DeleteBtn = ({id, setReload, setShow}) => {
+const DeleteBtn = ({ id, setReload, setShow }) => {
 
   const deleteRecipe = e => {
-    deleteReq(`https://recipes-homes-api.herokuapp.com/api/recipe/${id}`).then(data => {
-      console.log(data.$clusterTime.clusterTime);
-      setShow(data.$clusterTime.clusterTime);
-      setTimeout(() => setShow(false), 1200);
-      setReload(true);
-    });
-    
+    if (confirm('Are you sure you want to delete this recipe?')) {
+      deleteReq(`https://recipes-homes-api.herokuapp.com/api/recipe/${id}`).then(data => {
+        console.log(data.$clusterTime.clusterTime);
+        setShow(data.$clusterTime.clusterTime);
+        setTimeout(() => setShow(false), 1200);
+        setReload(true);
+      });
+    }
+
   };
 
   return (
